refactor: update infra test
